workflow-audio/requirements.txt
# Optimized Audio Service Requirements - Python 3.9 Compatible
# Core audio processing (essential) - FIXED VERSIONS

# TensorFlow MUST be installed first for essentia-tensorflow to work properly
tensorflow==2.8.4 # Required for essentia-tensorflow TensorFlow models

# Core audio processing - use official PyPI wheel with TensorFlow support
essentia-tensorflow  # Official PyPI wheel with TensorFlow models included
librosa==0.10.1
pydub==0.25.1
mutagen==1.47.0
soundfile==0.12.1

# Core FastAPI dependencies
fastapi==0.104.1  # Compatible with Python 3.9
uvicorn==0.24.0
pydantic==2.5.0  # More stable version
pydantic-settings==2.1.0
python-multipart==0.0.6

# Essential scientific computing - FIXED VERSIONS
numpy==1.24.3  # Critical: Must be <2.0.0 for Essentia
scipy==1.11.4

# ML for audio analysis
scikit-learn==1.3.2

# Data processing
pandas==2.1.4

# HTTP requests
requests==2.31.0

# Database and persistence - compatible versions
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis for caching
redis[hiredis]==5.0.1

# HTTP and async support
httpx==0.25.2
aiofiles==23.2.1

# Configuration and utilities
click==8.1.7
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Security and validation
email-validator==2.1.0
passlib[bcrypt]==1.7.4

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Type checking
mypy==1.7.1
types-redis==4.6.0.11